requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
Flask==3.0.2
Flask-Limiter==3.5.0 